requirements.txt
Pyrogram==2.0.106
TgCrypto==1.2.5
tinytuya==1.12.11
python-dotenv==1.0.0